+++ components/util/src/memfault_chunk_transport.c
@@ -20,6 +20,7 @@
 #include "memfault/util/chunk_transport.h"
 
 #include <stdbool.h>
+#include <string.h>
 
 #include "memfault/core/compiler.h"
 #include "memfault/core/math.h"
@@ -130,7 +131,23 @@ bool memfault_chunk_transport_get_next_chunk(sMfltChunkTransportCtx *ctx,
 
   ctx->read_msg(ctx->read_offset, &chunk_msg[chunk_msg_start_offset], bytes_to_read);
   ctx->read_offset += bytes_to_read;
-  *out_buf_len = chunk_msg_start_offset + bytes_to_read;
+  const size_t bytes_written = chunk_msg_start_offset + bytes_to_read;
+  const size_t buf_space_rem = *out_buf_len - bytes_written;
+  if (buf_space_rem != 0) {
+    // The encoded chunk consumes less space than the buffer provided. This can
+    // happen when we reach the end of the underlying message being encoded
+    //
+    // The Memfault backend allows for the last chunk in a message to exceed the
+    // size of the underlying message being sent. This can be useful for topologies
+    // that require sending chunks of a fixed size.
+    //
+    // Scrub the remaining part of the buffer for this situation with a know pattern
+    // for debug purposes so it's easier to visually see the end of the chunk and to
+    // prevent unintentional data from being sent.
+    memset(&chunk_msg[bytes_written], 0xBA, buf_space_rem);
+  }
+
+  *out_buf_len = bytes_written;
 
   return more_data;
 }